requirements.txt
yt-dlp>=2024.7.1
faster-whisper>=1.0.0
pydub>=0.25.1
PyYAML>=6.0.1
demucs>=4.0.0
torchcodec==0.8
Flask>=3.0.0
reportlab>=4.0.0
